requirements.txt
Django==5.2.3
djangorestframework==3.15.2
requests==2.32.3
python-decouple==3.8
gunicorn==21.2.0 